Update classification.md to correct a few function names
In the documented examples 2 functions are created to run the classification process: `classify` and `multi_classify`, however, instead of using these functions to run the process a generic `createClassification` function is called.

I presume this `createClassification` function name was inherited from the Python documentation.

Hope this helps to reduce potential confusion. Thanks for your work!